Split out tscredentials from nla.
