+++ libfreerdp/core/nla.c
@@ -49,6 +49,7 @@
 #include "../mit-krb5-pkinit/kinit.h"
 #include "nla.h"
 #include "smartcardlogon.h"
+#include "tscredentials.h"
 
 #define TAG FREERDP_TAG("core.nla")
 
@@ -72,113 +73,7 @@
 #define INIT_SECURITY_INTERFACE_NAME "InitSecurityInterfaceA"
 #endif
 
-
-/**
-* TSRequest ::= SEQUENCE {
-* 	version    [0] INTEGER,
-* 	negoTokens [1] NegoData OPTIONAL,
-* 	authInfo   [2] OCTET STRING OPTIONAL,
-* 	pubKeyAuth [3] OCTET STRING OPTIONAL,
-* 	errorCode  [4] INTEGER OPTIONAL
-* }
-*
-* NegoData ::= SEQUENCE OF NegoDataItem
-*
-* NegoDataItem ::= SEQUENCE {
-* 	negoToken [0] OCTET STRING
-* }
-*
-* TSCredentials ::= SEQUENCE {
-* 	cred_type    [0] INTEGER,
-* 	credentials [1] OCTET STRING
-* }
-*
-* TSPasswordCreds ::= SEQUENCE {
-* 	domainName  [0] OCTET STRING,
-* 	userName    [1] OCTET STRING,
-* 	password    [2] OCTET STRING
-* }
-*
-* TSSmartCardCreds ::= SEQUENCE {
-* 	pin        [0] OCTET STRING,
-* 	cspData    [1] TSCspDataDetail,
-* 	userHint   [2] OCTET STRING OPTIONAL,
-* 	domainHint [3] OCTET STRING OPTIONAL
-* }
-*
-* TSCspDataDetail ::= SEQUENCE {
-* 	keySpec       [0] INTEGER,
-* 	cardName      [1] OCTET STRING OPTIONAL,
-* 	readerName    [2] OCTET STRING OPTIONAL,
-* 	containerName [3] OCTET STRING OPTIONAL,
-* 	cspName       [4] OCTET STRING OPTIONAL
-* }
-*
-*
-* TSRemoteGuardCreds ::= SEQUENCE {
-*     logonCred         [0] TSRemoteGuardPackageCred,
-*     supplementalCreds [1] SEQUENCE OF TSRemoteGuardPackageCred OPTIONAL,
-* }
-*
-* TSRemoteGuardPackageCred ::=  SEQUENCE {
-*     packageName [0] OCTET STRING,
-*     credBuffer  [1] OCTET STRING,
-* }
-*
-*/
-
 #define NLA_PKG_NAME	NEGO_SSP_NAME
-
-
-typedef struct
-{
-	UINT32 KeySpec;
-	char* CardName;
-	char* ReaderName;
-	char* ContainerName;
-	char* CspName;
-} csp_data_detail;
-
-typedef struct
-{
-	char* Pin;
-	char* UserHint;   /* OPTIONAL */
-	char* DomainHint; /* OPTIONAL */
-	csp_data_detail* csp_data;
-} smartcard_creds;
-
-typedef struct
-{
-	char*    package_name;
-	unsigned credential_size;
-	BYTE*    credential;
-} remote_guard_package_cred;
-
-typedef struct
-{
-	remote_guard_package_cred*  login_cred;
-	unsigned                    supplemental_creds_count;
-	remote_guard_package_cred** supplemental_creds;
-} remote_guard_creds;
-
-typedef enum
-{
-	credential_type_password = 1,
-	credential_type_smartcard = 2,
-	credential_type_remote_guard = 6,
-	credential_type_default = credential_type_password,
-} credential_type;
-
-typedef struct
-{
-	credential_type          cred_type;
-	union
-	{
-		SEC_WINNT_AUTH_IDENTITY* password_creds;
-		smartcard_creds*         smartcard_creds;
-		remote_guard_creds*      remote_guard_creds;
-	} creds;
-} auth_identity;
 
 struct rdp_nla
 {
@@ -236,12 +131,6 @@ static SECURITY_STATUS nla_decrypt_public_key_echo(rdpNla* nla);
 static SECURITY_STATUS nla_decrypt_public_key_hash(rdpNla* nla);
 static SECURITY_STATUS nla_encrypt_ts_credentials(rdpNla* nla);
 static SECURITY_STATUS nla_decrypt_ts_credentials(rdpNla* nla);
-static BOOL nla_read_ts_password_creds(rdpNla* nla, wStream* s, size_t* length);
-static BOOL nla_read_ts_smartcard_creds(rdpNla* nla, wStream* s, size_t* length);
-static void SEC_WINNT_AUTH_IDENTITY_free(SEC_WINNT_AUTH_IDENTITY* identity);
-
-#define ber_sizeof_sequence_octet_string(length) ber_sizeof_contextual_tag(ber_sizeof_octet_string(length)) + ber_sizeof_octet_string(length)
-#define ber_write_sequence_octet_string(stream, context, value, length) ber_write_contextual_tag(stream, context, ber_sizeof_octet_string(length), TRUE) + ber_write_octet_string(stream, value, length)
 
 /* CredSSP Client-To-Server Binding Hash\0 */
 static const BYTE ClientServerHashMagic[] =
@@ -280,36 +169,7 @@ static const UINT32 NonceLength = 32;
 		}                                                       \
 	}while (0)
 
-static void memory_clear_and_free(void* memory, size_t size)
-{
-	if (memory)
-	{
-		memset(memory, 0, size);
-		free(memory);
-	}
-}
-
-static void* memdup(void* source, size_t size)
-{
-	void* destination = malloc(size);
-
-	if (destination != NULL)
-	{
-		memcpy(destination, source, size);
-	}
-
-	return destination;
-}
-
 /* ============================================================ */
-
-static void string_clear_and_free(char* string)
-{
-	if (string)
-	{
-		memory_clear_and_free(string, strlen(string));
-	}
-}
 
 
 /*
@@ -334,407 +194,6 @@ LPTSTR stringX_from_cstring(const char* cstring)
 }
 
 
-/* ============================================================ */
-
-/* SEC_WINNT_AUTH_IDENTITY contains only UTF-16 strings,  with length fields. */
-#define WSTRING_LENGTH_CLEAR_AND_FREE(structure, field)				\
-	memory_clear_and_free(structure->field, structure->field##Length * 2)
-#define WSTRING_LENGTH_SET_CSTRING(structure, field, cstring)				\
-	(structure->field##Length = (cstring						\
-	                             ?ConvertToUnicode(CP_UTF8, 0, cstring, -1, &(structure->field), 0)	\
-	                             :(structure->field = NULL, 0)))
-#define WSTRING_LENGTH_COPY(source, destination, field)				\
-	((destination->field##Length = source->field##Length),			\
-	 (destination->field = ((source->field == NULL)			\
-	                        ?NULL							\
-	                        :memdup(source->field, source->field##Length))))
-
-static SEC_WINNT_AUTH_IDENTITY* SEC_WINNT_AUTH_IDENTITY_new(char* user,  char* password,
-        char* domain)
-{
-	SEC_WINNT_AUTH_IDENTITY* password_creds;
-	CHECK_MEMORY(password_creds = malloc(sizeof(*password_creds)), NULL,
-	             "Could not allocate a SEC_WINNT_AUTH_IDENTITY structure");
-	password_creds->Flags = SEC_WINNT_AUTH_IDENTITY_UNICODE;
-	WSTRING_LENGTH_SET_CSTRING(password_creds, User, user);
-	WSTRING_LENGTH_SET_CSTRING(password_creds, Domain, domain);
-	WSTRING_LENGTH_SET_CSTRING(password_creds, Password, password);
-	return password_creds;
-}
-
-static SEC_WINNT_AUTH_IDENTITY* SEC_WINNT_AUTH_IDENTITY_deepcopy(SEC_WINNT_AUTH_IDENTITY* original)
-{
-	SEC_WINNT_AUTH_IDENTITY* copy;
-	CHECK_MEMORY(copy = calloc(1, sizeof(*copy)), NULL,
-	             "Could not allocate a SEC_WINNT_AUTH_IDENTITY structure");
-	copy->Flags = SEC_WINNT_AUTH_IDENTITY_UNICODE;
-	WSTRING_LENGTH_COPY(original, copy, User);
-	WSTRING_LENGTH_COPY(original, copy, Domain);
-	WSTRING_LENGTH_COPY(original, copy, Password);
-
-	if (((original->User != NULL) && (copy->User == NULL))
-	    || ((original->Password != NULL) && (copy->Password == NULL))
-	    || ((original->Domain != NULL) && (copy->Domain == NULL)))
-	{
-		SEC_WINNT_AUTH_IDENTITY_free(copy);
-		WLog_ERR(TAG, "%s:%d: %s() Could not allocate the fields of a SEC_WINNT_AUTH_IDENTITY structure",
-		         __FILE__, __LINE__, __FUNCTION__);
-		return NULL;
-	}
-
-	return copy;
-}
-
-static void SEC_WINNT_AUTH_IDENTITY_free(SEC_WINNT_AUTH_IDENTITY* password_creds)
-{
-	if (password_creds)
-	{
-		WSTRING_LENGTH_CLEAR_AND_FREE(password_creds, User);
-		WSTRING_LENGTH_CLEAR_AND_FREE(password_creds, Domain);
-		WSTRING_LENGTH_CLEAR_AND_FREE(password_creds, Password);
-	}
-
-	free(password_creds);
-}
-
-
-/* ============================================================ */
-
-static csp_data_detail* csp_data_detail_new_nocopy(UINT32 KeySpec,
-        char* CardName,
-        char* ReaderName,
-        char* ContainerName,
-        char* CspName)
-{
-	csp_data_detail* csp;
-	CHECK_MEMORY(csp = malloc(sizeof(*csp)), NULL, "Could not allocate a csp_data_detail structure");
-	csp->KeySpec = KeySpec;
-	csp->CardName = CardName;
-	csp->ReaderName = ReaderName;
-	csp->ContainerName = ContainerName;
-	csp->CspName = CspName;
-	return csp;
-}
-
-static csp_data_detail* csp_data_detail_new(UINT32 KeySpec,
-        char* cardname,
-        char* readername,
-        char* containername,
-        char* cspname)
-{
-	char* CardName = strdup(cardname);
-	char* ReaderName = strdup(readername);
-	char* ContainerName = strdup(containername);
-	char* CspName = strdup(cspname);
-
-	if (!CardName || !ReaderName || !ContainerName || !CspName)
-	{
-		free(CardName);
-		free(ReaderName);
-		free(ContainerName);
-		free(CspName);
-		WLog_ERR(TAG, "%s:%d: %s() Could not allocate CardName, ReaderName, ContainerName or CspName",
-		         __FILE__, __LINE__, __FUNCTION__);
-		return NULL;
-	}
-
-	return csp_data_detail_new_nocopy(KeySpec, CardName, ReaderName, ContainerName, CspName);
-}
-
-static csp_data_detail* csp_data_detail_deepcopy(csp_data_detail* original)
-{
-	return csp_data_detail_new(original->KeySpec,
-	                           original->CardName,
-	                           original->ReaderName,
-	                           original->ContainerName,
-	                           original->CspName);
-}
-
-static void csp_data_detail_free(csp_data_detail* csp)
-{
-	if (csp == NULL)
-	{
-		return;
-	}
-
-	string_clear_and_free(csp->CardName);
-	string_clear_and_free(csp->ReaderName);
-	string_clear_and_free(csp->ContainerName);
-	string_clear_and_free(csp->CspName);
-	memory_clear_and_free(csp, sizeof(*csp));
-}
-
-/* ============================================================ */
-
-static smartcard_creds* smartcard_creds_new_nocopy(char* Pin, char* UserHint, char* DomainHint,
-        csp_data_detail* csp_data)
-{
-	smartcard_creds* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a smartcard_creds structure");
-	that->Pin = Pin;
-	that->UserHint = UserHint;
-	that->DomainHint = DomainHint;
-	that->csp_data = csp_data;
-	return that;
-}
-
-static smartcard_creds* smartcard_creds_new(char* pin, char* userhint, char* domainhint,
-        csp_data_detail* cspdata)
-{
-	char* Pin = strdup(pin);
-	char* UserHint = strdup(userhint);
-	char* DomainHint = strdup(domainhint);
-	csp_data_detail* cspData = csp_data_detail_deepcopy(cspdata);
-
-	if ((Pin == NULL) || (UserHint == NULL) || (DomainHint == NULL) || (cspData == NULL))
-	{
-		free(Pin);
-		free(UserHint);
-		free(DomainHint);
-		csp_data_detail_free(cspData);
-		WLog_ERR(TAG,
-		         "%s:%d: %s() Could not allocate Pin, UserHint or DomainHint,  or copy CSP Data Details",
-		         __FILE__, __LINE__, __FUNCTION__);
-		return NULL;
-	}
-
-	return smartcard_creds_new_nocopy(Pin, UserHint, DomainHint, cspData);
-}
-
-static smartcard_creds* smartcard_creds_deepcopy(smartcard_creds* original)
-{
-	return smartcard_creds_new(original->Pin, original->UserHint, original->DomainHint,
-	                           original->csp_data);
-}
-
-static void smartcard_creds_free(smartcard_creds* that)
-{
-	if (that == NULL)
-	{
-		return;
-	}
-
-	string_clear_and_free(that->Pin);
-	string_clear_and_free(that->UserHint);
-	string_clear_and_free(that->DomainHint);
-	csp_data_detail_free(that->csp_data);
-	memory_clear_and_free(that, sizeof(*that));
-}
-
-/* ============================================================ */
-
-static remote_guard_package_cred* remote_guard_package_cred_new_nocopy(char* package_name,
-        unsigned credential_size,
-        BYTE*    credential)
-{
-	remote_guard_package_cred* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a remote_guard_package_cred structure");
-	that->package_name = package_name;
-	that->credential_size = credential_size;
-	that->credential = credential;
-	return that;
-}
-
-static void remote_guard_package_cred_deepfree(remote_guard_package_cred* that)
-{
-	if (that == NULL)
-	{
-		return;
-	}
-
-	free(that->package_name);
-	free(that->credential);
-	memset(that, 0, sizeof(*that));
-	free(that);
-}
-
-static remote_guard_package_cred* remote_guard_package_cred_deepcopy(remote_guard_package_cred*
-        that)
-{
-	remote_guard_package_cred* copy;
-	CHECK_MEMORY(copy = malloc(sizeof(*copy)), NULL,
-	             "Could not allocate a remote_guard_package_cred structure");
-	copy->package_name = strdup(that->package_name);
-	copy->credential_size = that->credential_size;
-	copy->credential = memdup(that->credential, that->credential_size);
-
-	if ((copy->package_name == NULL) || (copy->credential == NULL))
-	{
-		remote_guard_package_cred_deepfree(copy);
-		CHECK_MEMORY(NULL, NULL, "Could not allocate a field of a remote_guard_package_cred structure");
-	}
-
-	return copy;
-}
-
-static void remote_guard_creds_free(remote_guard_creds* that);
-static remote_guard_creds* remote_guard_creds_new_nocopy(char* package_name,
-        unsigned credenial_size,
-        BYTE*    credential)
-{
-	remote_guard_creds* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a remote_guard_creds structure");
-	that->login_cred = remote_guard_package_cred_new_nocopy(package_name,
-	                   credenial_size, credential);
-	that->supplemental_creds_count = 0;
-	that->supplemental_creds = NULL;
-
-	if (that->login_cred == NULL)
-	{
-		remote_guard_creds_free(that);
-		that = NULL;
-	}
-
-	return that;
-}
-
-static remote_guard_creds* remote_guard_creds_deepcopy(remote_guard_creds* that)
-{
-	unsigned i;
-	remote_guard_creds* copy;
-	CHECK_MEMORY(copy = malloc(sizeof(*copy)), NULL,
-	             "Could not allocate a remote_guard_creds structure");
-	copy->login_cred = remote_guard_package_cred_deepcopy(that->login_cred);
-	copy->supplemental_creds_count = that->supplemental_creds_count;
-	copy->supplemental_creds = malloc(copy->supplemental_creds_count * sizeof(
-	                                      copy->supplemental_creds[0]));
-
-	for (i = 0; i < copy->supplemental_creds_count; i++)
-	{
-		copy->supplemental_creds[i] = remote_guard_package_cred_deepcopy(that->supplemental_creds[i]);
-
-		if (copy->supplemental_creds[i] == NULL)
-		{
-			break;
-		}
-	}
-
-	if ((copy->login_cred == NULL) || (copy->supplemental_creds == NULL)
-	    || (i < copy->supplemental_creds_count))
-	{
-		remote_guard_creds_free(copy);
-		CHECK_MEMORY(NULL, NULL, "Could not allocate the a field of a remote_guard_creds structure");
-	}
-
-	return copy;
-}
-
-static void remote_guard_creds_free(remote_guard_creds* that)
-{
-	remote_guard_package_cred_deepfree(that->login_cred);
-
-	if (that->supplemental_creds != NULL)
-	{
-		unsigned i;
-
-		for (i = 0; i < that->supplemental_creds_count; i++)
-		{
-			remote_guard_package_cred_deepfree(that->supplemental_creds[i]);
-		}
-
-		memory_clear_and_free(that->supplemental_creds,
-		                      that->supplemental_creds_count * sizeof(that->supplemental_creds[0]));
-	}
-
-	memory_clear_and_free(that, sizeof(* that));
-}
-
-/* ============================================================ */
-
-static auth_identity* auth_identity_new_password(SEC_WINNT_AUTH_IDENTITY* password_creds)
-{
-	auth_identity* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a auth_identity structure");
-	that->cred_type = credential_type_password;
-	that->creds.password_creds = password_creds;
-	return that;
-}
-
-static auth_identity* auth_identity_new_smartcard(smartcard_creds* smartcard_creds)
-{
-	auth_identity* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a auth_identity structure");
-	that->cred_type = credential_type_smartcard;
-	that->creds.smartcard_creds = smartcard_creds;
-	return that;
-}
-
-static auth_identity* auth_identity_new_remote_guard(remote_guard_creds* remote_guard_creds)
-{
-	auth_identity* that;
-	CHECK_MEMORY(that = malloc(sizeof(*that)), NULL,
-	             "Could not allocate a auth_identity structure");
-	that->cred_type = credential_type_remote_guard;
-	that->creds.remote_guard_creds = remote_guard_creds;
-	return that;
-}
-
-static void auth_identity_free(auth_identity* that)
-{
-	if (that != NULL)
-	{
-		switch (that->cred_type)
-		{
-			case credential_type_password:
-				SEC_WINNT_AUTH_IDENTITY_free(that->creds.password_creds);
-				break;
-
-			case credential_type_smartcard:
-				smartcard_creds_free(that->creds.smartcard_creds);
-				break;
-
-			case credential_type_remote_guard:
-				remote_guard_creds_free(that->creds.remote_guard_creds);
-				break;
-		}
-
-		memory_clear_and_free(that, sizeof(*that));
-	}
-}
-
-static auth_identity* auth_identity_deepcopy(auth_identity* that)
-{
-#define CHECK_COPY(field, copier)					\
-	if (that->creds.field != NULL)					\
-	{								\
-		if ((field = copier(that->creds.field)) == NULL)	\
-		{							\
-			goto failure;					\
-		}							\
-	}
-	SEC_WINNT_AUTH_IDENTITY* password_creds = NULL;
-	smartcard_creds* smartcard_creds = NULL;
-	remote_guard_creds* remote_guard_creds = NULL;
-
-	switch (that->cred_type)
-	{
-		case credential_type_password:
-			CHECK_COPY(password_creds, SEC_WINNT_AUTH_IDENTITY_deepcopy);
-			return auth_identity_new_password(password_creds);
-
-		case credential_type_smartcard:
-			CHECK_COPY(smartcard_creds, smartcard_creds_deepcopy);
-			return auth_identity_new_smartcard(smartcard_creds);
-
-		case credential_type_remote_guard:
-			CHECK_COPY(remote_guard_creds, remote_guard_creds_deepcopy);
-			return auth_identity_new_remote_guard(remote_guard_creds);
-	}
-
-failure:
-	SEC_WINNT_AUTH_IDENTITY_free(password_creds);
-	smartcard_creds_free(smartcard_creds);
-	remote_guard_creds_free(remote_guard_creds);
-	return NULL;
-#undef CHECK_COPY
-}
-
-/* ============================================================ */
 
 /**
 * Returns whether the username is found in the SAM database.
@@ -2301,567 +1760,6 @@ fail:
 	return status;
 }
 
-static size_t nla_sizeof_ts_password_creds(SEC_WINNT_AUTH_IDENTITY* password_creds)
-{
-	size_t length = 0;
-
-	if (password_creds)
-	{
-		length += ber_sizeof_sequence_octet_string(password_creds->DomainLength * 2);
-		length += ber_sizeof_sequence_octet_string(password_creds->UserLength * 2);
-		length += ber_sizeof_sequence_octet_string(password_creds->PasswordLength * 2);
-	}
-
-	return length;
-}
-
-
-/*
-string_length returns the number of characters in the utf-8 string.
-That is, the number of WCHAR in the string converted to "unicode".
-*/
-static int string_length(char* string)
-{
-	if (string == NULL)
-	{
-		return 0;
-	}
-	else
-	{
-		LPWSTR wstring = NULL;
-		int length =  ConvertToUnicode(CP_UTF8, 0, string, -1, & wstring, 0);
-		free(wstring);
-		return (length < 0) ? 0 : length;
-	}
-}
-
-static int nla_sizeof_ts_cspdatadetail(csp_data_detail*  csp_data)
-{
-	int length = 0;
-
-	if (csp_data)
-	{
-		length += ber_sizeof_contextual_tag(ber_sizeof_integer(csp_data->KeySpec));
-		length += ber_sizeof_integer(csp_data->KeySpec);
-		length += ber_sizeof_sequence_octet_string(string_length(csp_data->CardName) * 2);
-		length += ber_sizeof_sequence_octet_string(string_length(csp_data->ReaderName) * 2);
-		length += ber_sizeof_sequence_octet_string(string_length(csp_data->ContainerName) * 2);
-		length += ber_sizeof_sequence_octet_string(string_length(csp_data->CspName) * 2);
-	}
-
-	return length;
-}
-
-static int nla_sizeof_sequence_ts_cspdatadetail(csp_data_detail*  csp_data)
-{
-	int length = 0;
-	length += nla_sizeof_ts_cspdatadetail(csp_data);
-	length += ber_sizeof_sequence_tag(length);
-	length += ber_sizeof_contextual_tag(length);
-	return length;
-}
-
-static int nla_sizeof_ts_smartcard_creds(auth_identity* identity)
-{
-	int length = 0;
-	length += ber_sizeof_sequence_octet_string(string_length(identity->creds.smartcard_creds->Pin) * 2);
-	length += nla_sizeof_sequence_ts_cspdatadetail(identity->creds.smartcard_creds->csp_data);
-	length += ber_sizeof_sequence_octet_string(string_length(identity->creds.smartcard_creds->UserHint)
-	          * 2);
-	length += ber_sizeof_sequence_octet_string(string_length(
-	              identity->creds.smartcard_creds->DomainHint) * 2);
-	return length;
-}
-
-static int nla_sizeof_ts_remote_guard_package_cred(remote_guard_package_cred* package_cred)
-{
-	int length = 0;
-	length += ber_sizeof_sequence_octet_string(string_length(package_cred->package_name) * 2);
-	length += ber_sizeof_sequence_octet_string(package_cred->credential_size);
-	return length;
-}
-
-static int nla_sizeof_sequence_ts_remote_guard_package_cred(remote_guard_package_cred* package_cred)
-{
-	int length = 0;
-	length += nla_sizeof_ts_remote_guard_package_cred(package_cred);
-	length += ber_sizeof_sequence_tag(length);
-	length += ber_sizeof_contextual_tag(length);
-	return length;
-}
-
-static int nla_sizeof_ts_remote_guard_creds(auth_identity* identity)
-{
-	int length = 0;
-	unsigned i;
-	length += nla_sizeof_sequence_ts_remote_guard_package_cred(
-	              identity->creds.remote_guard_creds->login_cred);
-	length += ber_sizeof_sequence_tag(length);
-	length += ber_sizeof_contextual_tag(length);
-
-	for (i = 0; i < identity->creds.remote_guard_creds->supplemental_creds_count; i++)
-	{
-		length += nla_sizeof_sequence_ts_remote_guard_package_cred(
-		              identity->creds.remote_guard_creds->supplemental_creds[i]);
-	}
-
-	return length;
-}
-
-static int nla_sizeof_ts_creds(auth_identity* identity)
-{
-	switch (identity->cred_type)
-	{
-		case credential_type_password:
-			return nla_sizeof_ts_password_creds(identity->creds.password_creds);
-
-		case credential_type_smartcard:
-			return nla_sizeof_ts_smartcard_creds(identity);
-
-		case credential_type_remote_guard:
-			return nla_sizeof_ts_remote_guard_creds(identity);
-
-		default:
-			return 0;
-	}
-}
-
-static size_t nla_sizeof_ts_credentials(auth_identity*   identity)
-{
-	size_t size = 0;
-	size += ber_sizeof_integer(1);
-	size += ber_sizeof_contextual_tag(ber_sizeof_integer(1));
-	size += ber_sizeof_sequence_octet_string(ber_sizeof_sequence(nla_sizeof_ts_creds(identity)));
-	return size;
-}
-
-
-static BOOL nla_read_octet_string_field(rdpNla* nla, wStream* s, const char* field_name,
-                                        int contextual_tag,
-                                        UINT16** field, UINT32* field_length)
-{
-	size_t length = 0;
-
-	if (!ber_read_contextual_tag(s, contextual_tag, &length, TRUE) ||
-	    !ber_read_octet_string_tag(s, &length))
-	{
-		return FALSE;
-	}
-
-	if (length == 0)
-	{
-		CHECK_MEMORY((*field) = calloc(1, 2),
-		             FALSE, "Cannot allocate %s", field_name);
-	}
-	else
-	{
-		CHECK_MEMORY((*field) = malloc(length),
-		             FALSE, "Cannot allocate %s", field_name);
-		CopyMemory((*field), Stream_Pointer(s), length);
-		Stream_Seek(s, length);
-	}
-
-	(*field_length) = (UINT32) length / 2;
-	return TRUE;
-}
-
-static BOOL nla_read_octet_string_string(rdpNla* nla, wStream* s, const char* field_name,
-        int contextual_tag,
-        char** string)
-{
-	int result;
-	UINT16* field = NULL;
-	UINT32 field_length = 0;
-	(*string) = NULL;
-
-	if (!nla_read_octet_string_field(nla, s, field_name, contextual_tag, &field, &field_length))
-	{
-		return 0;
-	}
-
-	result = ConvertFromUnicode(CP_UTF8, 0, field, 2 * field_length, string, 0, 0, FALSE);
-	free(field);
-	return 0 <= result;
-}
-
-static int nla_write_octet_string_string(wStream* s, const char* field_name, int contextual_tag,
-        char* string)
-{
-	LPWSTR buffer = NULL;
-	int length = ConvertToUnicode(CP_UTF8, 0, string, -1, &buffer, 0);
-
-	if (length < 0)
-	{
-		WLog_ERR(TAG, "Cannot ConvertToUnicode '%s'", string);
-		return 0;
-	}
-
-	int increment = ber_write_sequence_octet_string(s, contextual_tag, (BYTE*)buffer, length * 2);
-	free(buffer);
-	return increment;
-}
-
-
-static BOOL nla_read_ts_password_creds(rdpNla* nla, wStream* s, size_t* length)
-{
-	if (!nla->identity)
-	{
-		WLog_ERR(TAG, "nla->identity is NULL!");
-		return FALSE;
-	}
-
-	/* TSPasswordCreds (SEQUENCE)
-	* Initialise to default values. */
-	SEC_WINNT_AUTH_IDENTITY_free(nla->identity->creds.password_creds);
-	nla->identity->creds.password_creds = SEC_WINNT_AUTH_IDENTITY_new(NULL, NULL, NULL);
-
-	if (nla->identity->creds.password_creds == NULL)
-	{
-		return FALSE;
-	}
-
-	if (!ber_read_sequence_tag(s, length))
-		return FALSE;
-
-	/* The sequence is empty, return early,
-	* TSPasswordCreds (SEQUENCE) is optional. */
-	if ((*length) == 0)
-		return TRUE;
-
-	return (nla_read_octet_string_field(nla, s, "[0] domainName (OCTET STRING)", 0,
-	                                    &nla->identity->creds.password_creds->Domain,
-	                                    &nla->identity->creds.password_creds->DomainLength) &&
-	        nla_read_octet_string_field(nla, s, "[1] userName (OCTET STRING)", 1,
-	                                    &nla->identity->creds.password_creds->User,
-	                                    &nla->identity->creds.password_creds->UserLength) &&
-	        nla_read_octet_string_field(nla, s, "[2] password (OCTET STRING)", 2,
-	                                    &nla->identity->creds.password_creds->Password,
-	                                    &nla->identity->creds.password_creds->PasswordLength));
-}
-
-static BOOL nla_read_ts_cspdatadetail(rdpNla* nla, wStream* s, size_t* length)
-{
-	if (!nla->identity->creds.smartcard_creds->csp_data)
-	{
-		return FALSE;
-	}
-
-	/* TSCspDataDetail (SEQUENCE)
-	* Initialise to default values. */
-	csp_data_detail_free(nla->identity->creds.smartcard_creds->csp_data);
-	nla->identity->creds.smartcard_creds->csp_data = csp_data_detail_new_nocopy(0, NULL, NULL, NULL,
-	        NULL);
-
-	if (nla->identity->creds.smartcard_creds->csp_data == NULL)
-	{
-		return FALSE;
-	}
-
-	if (!ber_read_sequence_tag(s, length))
-	{
-		return FALSE;
-	}
-
-	/* The sequence is empty, return early,
-	* TSCspDataDetail (SEQUENCE) is optional. */
-	if (*length == 0)
-		return TRUE;
-
-	/* [0] keySpec (INTEGER) */
-	if (!ber_read_contextual_tag(s, 0, length, TRUE) ||
-	    !ber_read_integer(s, &nla->identity->creds.smartcard_creds->csp_data->KeySpec))
-	{
-		return FALSE;
-	}
-
-	return (nla_read_octet_string_string(nla, s, "[1] cardName (OCTET STRING)", 1,
-	                                     &nla->identity->creds.smartcard_creds->csp_data->CardName) &&
-	        nla_read_octet_string_string(nla, s, "[2] readerName (OCTET STRING)", 2,
-	                                     &nla->identity->creds.smartcard_creds->csp_data->ReaderName) &&
-	        nla_read_octet_string_string(nla, s, "[3] containerName (OCTET STRING)", 3,
-	                                     &nla->identity->creds.smartcard_creds->csp_data->ContainerName) &&
-	        nla_read_octet_string_string(nla, s, "[4] cspName (OCTET STRING)", 4,
-	                                     &nla->identity->creds.smartcard_creds->csp_data->CspName));
-}
-
-static BOOL nla_read_ts_smartcard_creds(rdpNla* nla, wStream* s, size_t* length)
-{
-	if ((nla->identity->cred_type != credential_type_smartcard)
-	    || (nla->identity->creds.smartcard_creds == NULL))
-	{
-		return FALSE;
-	}
-
-	/* TSSmartCardCreds (SEQUENCE)
-	* Initialize to default values. */
-	smartcard_creds_free(nla->identity->creds.smartcard_creds);
-	nla->identity->creds.smartcard_creds = smartcard_creds_new_nocopy(NULL, NULL, NULL,
-	                                       csp_data_detail_new_nocopy(0, NULL, NULL, NULL, NULL));
-
-	if (nla->identity->creds.smartcard_creds == NULL)
-	{
-		return FALSE;
-	}
-
-	if (!ber_read_sequence_tag(s, length))
-		return FALSE;
-
-	/* The sequence is empty, return early,
-	* TSSmartCardCreds (SEQUENCE) is optional. */
-	if ((*length) == 0)
-		return TRUE;
-
-	return (nla_read_octet_string_string(nla, s, "[0] Pin (OCTET STRING)", 0,
-	                                     &nla->identity->creds.smartcard_creds->Pin) &&
-	        nla_read_ts_cspdatadetail(nla, s, length) &&
-	        nla_read_octet_string_string(nla, s, "[2] UserHint (OCTET STRING)", 2,
-	                                     &nla->identity->creds.smartcard_creds->UserHint) &&
-	        nla_read_octet_string_string(nla, s, "[3] DomainHint (OCTET STRING", 3,
-	                                     &nla->identity->creds.smartcard_creds->DomainHint));
-}
-
-static size_t nla_write_ts_password_creds(SEC_WINNT_AUTH_IDENTITY*   password_creds, wStream* s)
-{
-	size_t size = 0;
-	size_t innerSize = nla_sizeof_ts_password_creds(password_creds);
-	/* TSPasswordCreds (SEQUENCE) */
-	size += ber_write_sequence_tag(s, innerSize);
-
-	if (password_creds)
-	{
-		/* [0] domainName (OCTET STRING) */
-		size += ber_write_sequence_octet_string(
-		            s, 0, (BYTE*) password_creds->Domain,
-		            password_creds->DomainLength * 2);
-		/* [1] userName (OCTET STRING) */
-		size += ber_write_sequence_octet_string(
-		            s, 1, (BYTE*) password_creds->User,
-		            password_creds->UserLength * 2);
-		/* [2] password (OCTET STRING) */
-		size += ber_write_sequence_octet_string(
-		            s, 2, (BYTE*) password_creds->Password,
-		            password_creds->PasswordLength * 2);
-	}
-
-	return size;
-}
-
-static int nla_write_ts_csp_data_detail(csp_data_detail*   csp_data, wStream* s, int contextual_tag)
-{
-	int size = 0;
-
-	if (csp_data != NULL)
-	{
-		size += ber_write_contextual_tag(s, 1, ber_sizeof_sequence(nla_sizeof_ts_cspdatadetail(csp_data)),
-		                                 TRUE);
-		size += ber_write_sequence_tag(s, nla_sizeof_ts_cspdatadetail(csp_data));
-		/* [0] KeySpec (INTEGER) */
-		size += ber_write_contextual_tag(s, 0, ber_sizeof_integer(csp_data->KeySpec), TRUE);
-		size += ber_write_integer(s, csp_data->KeySpec);
-		size += nla_write_octet_string_string(s, "[1] CardName (OCTET STRING)", 1, csp_data->CardName);
-		size += nla_write_octet_string_string(s, "[2] ReaderName (OCTET STRING)", 2, csp_data->ReaderName);
-		size += nla_write_octet_string_string(s, "[3] ContainerName (OCTET STRING)", 3,
-		                                      csp_data->ContainerName);
-		size += nla_write_octet_string_string(s, "[4] CspName (OCTET STRING)", 4, csp_data->CspName);
-	}
-
-	return size;
-}
-
-static int nla_write_ts_smartcard_creds(auth_identity* identity, wStream* s)
-{
-	int size = 0;
-	int innerSize = nla_sizeof_ts_smartcard_creds(identity);
-	/* TSSmartCardCreds (SEQUENCE) */
-	size += ber_write_sequence_tag(s, innerSize);
-
-	if ((identity != NULL) && (identity->creds.smartcard_creds != NULL))
-	{
-		size += nla_write_octet_string_string(s, "[0] Pin (OCTET STRING)", 0,
-		                                      identity->creds.smartcard_creds->Pin);
-		size += nla_write_ts_csp_data_detail(identity->creds.smartcard_creds->csp_data, s,
-		                                     1); /* [1] CspDataDetail (TSCspDataDetail) (SEQUENCE) */
-		size += nla_write_octet_string_string(s, "[2] userHint (OCTET STRING)", 2,
-		                                      identity->creds.smartcard_creds->UserHint);
-		size += nla_write_octet_string_string(s, "[3] domainHint (OCTET STRING)", 3,
-		                                      identity->creds.smartcard_creds->DomainHint);
-	}
-
-	return size;
-}
-
-static BOOL nla_read_ts_remote_guard_creds(rdpNla* nla, wStream* s, size_t* length)
-{
-	/*
-	TODO
-	if (!nla->identity->creds.smartcard_creds)
-	{
-		return FALSE;
-	}
-	*/
-	/* TSRemoteGuardCreds (SEQUENCE)
-	* Initialize to default values. */
-	/* smartcard_creds_free(nla->identity->creds.smartcard_creds); */
-	/* nla->identity->creds.smartcard_creds = smartcard_creds_new_nocopy(NULL, NULL, NULL); */
-	/*  */
-	/* if (nla->identity->creds.smartcard_creds == NULL) */
-	/* { */
-	/* 	return FALSE; */
-	/* } */
-	/*  */
-	/* if (!ber_read_sequence_tag(s, length)) */
-	/* 	return FALSE; */
-	/* The sequence is empty, return early,
-	* TSSmartCardCreds (SEQUENCE) is optional. */
-	/* if ((*length) == 0) */
-	/* 	return TRUE; */
-	/*  */
-	/* return (nla_read_octet_string_string(nla, s, "[0] Pin (OCTET STRING)", 0, */
-	/*                                      &nla->identity->creds.smartcard_creds->Pin) && */
-	/*         nla_read_ts_cspdatadetail(nla, s, length) && */
-	/*         nla_read_octet_string_string(nla, s, "[2] UserHint (OCTET STRING)", 2, */
-	/*                                      &nla->identity->creds.smartcard_creds->UserHint) && */
-	/*         nla_read_octet_string_string(nla, s, "[3] DomainHint (OCTET STRING", 3, */
-	/*                                      &nla->identity->creds.smartcard_creds->DomainHint)); */
-	return FALSE;
-}
-
-static int nla_write_ts_remote_guard_package_cred(remote_guard_package_cred* package_cred,
-        wStream* s)
-{
-	int size = 0;
-
-	if (package_cred != NULL)
-	{
-		size += nla_write_octet_string_string(s, "packageName [0] OCTET STRING",
-		                                      0, package_cred->package_name);
-		/* credBuffer  [1] OCTET STRING, */
-		size += ber_write_sequence_octet_string(s, 1, package_cred->credential,
-		                                        package_cred->credential_size);
-	}
-
-	return size;
-}
-
-static int nla_write_ts_remote_guard_creds(auth_identity* identity, wStream* s)
-{
-	unsigned i;
-	int size = 0;
-	int innerSize = nla_sizeof_ts_remote_guard_creds(identity);
-	/* TSRemoteGuardCreds (SEQUENCE) */
-	size += ber_write_sequence_tag(s, innerSize);
-	size += nla_write_ts_remote_guard_package_cred(identity->creds.remote_guard_creds->login_cred, s);
-	/* TODO: add sequence header? */
-	/* credSize = ber_sizeof_sequence(nla_sizeof_ts_creds(identity)); */
-	/* size += ber_write_contextual_tag(s, 1, ber_sizeof_octet_string(credSize), TRUE); */
-
-	for (i = 0; i < identity->creds.remote_guard_creds->supplemental_creds_count; i++)
-	{
-		size += nla_write_ts_remote_guard_package_cred(
-		            identity->creds.remote_guard_creds->supplemental_creds[i], s);
-	}
-
-	/* if ((identity != NULL) && (identity->creds.smartcard_creds != NULL)) */
-	/* { */
-	/* size += nla_write_octet_string_string(s, "[0] Pin (OCTET STRING)", 0, */
-	/*                                       identity->creds.smartcard_creds->Pin); */
-	/* size += nla_write_ts_csp_data_detail(identity->creds.csp_data, s, */
-	/*                                      1); /\* [1] CspDataDetail (TSCspDataDetail) (SEQUENCE) *\/ */
-	/* size += nla_write_octet_string_string(s, "[2] userHint (OCTET STRING)", 2, */
-	/*                                       identity->creds.smartcard_creds->UserHint); */
-	/* size += nla_write_octet_string_string(s, "[3] domainHint (OCTET STRING)", 3, */
-	/*                                       identity->creds.smartcard_creds->DomainHint); */
-	/* } */
-	return size;
-}
-
-static int nla_read_ts_creds(rdpNla* nla, wStream* s, credential_type cred_type, size_t* length)
-{
-	switch (cred_type)
-	{
-		case credential_type_password:
-			return nla_read_ts_password_creds(nla, s, length);
-
-		case credential_type_smartcard:
-			return nla_read_ts_smartcard_creds(nla, s, length);
-
-		case credential_type_remote_guard:
-			return nla_read_ts_remote_guard_creds(nla, s, length);
-
-		default:
-			WLog_ERR(TAG,  "cred_type unknown: %d\n", cred_type);
-			return 0;
-	}
-}
-
-static int nla_write_ts_creds(auth_identity* identity, wStream* s)
-{
-	switch (identity->cred_type)
-	{
-		case credential_type_password:
-			return nla_write_ts_password_creds(identity->creds.password_creds, s);
-
-		case credential_type_smartcard:
-			return nla_write_ts_smartcard_creds(identity, s);
-
-		case credential_type_remote_guard:
-			return nla_write_ts_remote_guard_creds(identity, s);
-
-		default:
-			WLog_ERR(TAG,  "cred_type unknown: %d\n", identity->cred_type);
-			return 0;
-	}
-}
-
-static BOOL nla_read_ts_credentials(rdpNla* nla, PSecBuffer ts_credentials)
-{
-	wStream* s;
-	size_t length = 0;
-	size_t ts_creds_length = 0;
-	UINT32* value = NULL;
-	BOOL ret;
-
-	if (!ts_credentials || !ts_credentials->pvBuffer)
-		return FALSE;
-
-	s = Stream_New(ts_credentials->pvBuffer, ts_credentials->cbBuffer);
-
-	if (!s)
-	{
-		WLog_ERR(TAG, "Stream_New failed!");
-		return FALSE;
-	}
-
-	/* TSCredentials (SEQUENCE) */
-	ret = (ber_read_sequence_tag(s, &length) &&
-	       /* [0] credType (INTEGER) */
-	       ber_read_contextual_tag(s, 0, &length, TRUE) &&
-	       ber_read_integer(s, value) &&
-	       /* [1] credentials (OCTET STRING) */
-	       ber_read_contextual_tag(s, 1, &length, TRUE) &&
-	       ber_read_octet_string_tag(s, &ts_creds_length) &&
-	       nla_read_ts_creds(nla, s, *value, &length)) ;
-	Stream_Free(s, FALSE);
-	return ret;
-}
-
-static size_t nla_write_ts_credentials(rdpNla* nla, wStream* s, auth_identity* identity)
-{
-	int size = 0;
-	int credSize = 0;
-	int innerSize = nla_sizeof_ts_credentials(identity);
-	/* TSCredentials (SEQUENCE) */
-	size += ber_write_sequence_tag(s, innerSize);
-	/* [0] credType (INTEGER) */
-	size += ber_write_contextual_tag(s, 0, ber_sizeof_integer(identity->cred_type), TRUE);
-	size += ber_write_integer(s, identity->cred_type);
-	/* [1] credentials (OCTET STRING) */
-	credSize = ber_sizeof_sequence(nla_sizeof_ts_creds(identity));
-	size += ber_write_contextual_tag(s, 1, ber_sizeof_octet_string(credSize), TRUE);
-	size += ber_write_octet_string_tag(s, credSize);
-	size += nla_write_ts_creds(identity, s);
-	return size;
-}
 
 /**
 * Encode TSCredentials structure.
@@ -2907,7 +1805,7 @@ static BOOL nla_encode_ts_credentials(rdpNla* nla)
 		goto cleanup;
 	}
 
-	nla_write_ts_credentials(nla, s, identity);
+	nla_write_ts_credentials(nla->identity, s);
 	result = TRUE;
 cleanup:
 
@@ -3049,23 +1947,44 @@ static SECURITY_STATUS nla_decrypt_ts_credentials(rdpNla* nla)
 	Message.pBuffers = (PSecBuffer) &Buffers;
 	status = nla->table->DecryptMessage(&nla->context, &Message, nla->recvSeqNum++, &pfQOP);
 
-	if (status != SEC_E_OK)
+	if (status == SEC_E_OK)
 	{
-		WLog_ERR(TAG, "DecryptMessage failure %s [0x%08"PRIX32"]",
-		         GetSecurityStatusString(status), status);
+		auth_identity* identity = nla_read_ts_credentials(&Buffers[1]);
+
+		if (identity == NULL)
+		{
+			free(buffer);
+			return SEC_E_INSUFFICIENT_MEMORY;
+		}
+
+		auth_identity_free(nla->identity);
+		nla->identity = identity;
 		free(buffer);
-		return status;
+		return SEC_E_OK;
 	}
 
-	if (!nla_read_ts_credentials(nla, &Buffers[1]))
-	{
-		free(buffer);
-		return SEC_E_INSUFFICIENT_MEMORY;
-	}
-
+	WLog_ERR(TAG, "DecryptMessage failure %s [0x%08"PRIX32"]",
+	         GetSecurityStatusString(status), status);
 	free(buffer);
-	return SEC_E_OK;
+	return status;
 }
+
+/**
+* TSRequest ::= SEQUENCE {
+* 	version    [0] INTEGER,
+* 	negoTokens [1] NegoData OPTIONAL,
+* 	authInfo   [2] OCTET STRING OPTIONAL,
+* 	pubKeyAuth [3] OCTET STRING OPTIONAL,
+* 	errorCode  [4] INTEGER OPTIONAL
+* }
+*
+* NegoData ::= SEQUENCE OF NegoDataItem
+*
+* NegoDataItem ::= SEQUENCE {
+* 	negoToken [0] OCTET STRING
+* }
+*
+*/
 
 static size_t nla_sizeof_nego_token(size_t length)
 {
