+++ libfreerdp/core/test/TestTSRequestRemoteGuardCreds.c
@@ -1,0 +1,32 @@
+
+
+BYTE result[] = { 0x30, 0x81, 0x86, 0xa0, 0x2d, 0x80, 0x11, 0x4d, 0x79, 0x53, 0x65, 0x63, 0x75, 0x72, 0x69, 0x74,
+                  0x79, 0x50, 0x61, 0x63, 0x6b, 0x61, 0x67, 0x65, 0x81, 0x18, 0x4d, 0x79, 0x20, 0x42, 0x72, 0x65,
+                  0x61, 0x74, 0x68, 0x20, 0x49, 0x73, 0x20, 0x4d, 0x79, 0x20, 0x50, 0x61, 0x73, 0x73, 0x77, 0x6f,
+                  0x72, 0x64, 0xa1, 0x55, 0x30, 0x3c, 0x80, 0x1a, 0x41, 0x6c, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x74,
+                  0x69, 0x76, 0x65, 0x53, 0x65, 0x63, 0x75, 0x72, 0x69, 0x74, 0x79, 0x50, 0x61, 0x63, 0x6b, 0x61,
+                  0x67, 0x65, 0x81, 0x1e, 0x4d, 0x79, 0x20, 0x42, 0x72, 0x65, 0x61, 0x74, 0x68, 0x20, 0x49, 0x73,
+                  0x20, 0x4d, 0x79, 0x20, 0x4f, 0x74, 0x68, 0x65, 0x72, 0x20, 0x50, 0x61, 0x73, 0x73, 0x77, 0x6f,
+                  0x72, 0x64, 0x30, 0x15, 0x80, 0x03, 0x50, 0x41, 0x4d, 0x81, 0x0e, 0x66, 0x6f, 0x6f, 0x62, 0x61,
+                  0x72, 0x62, 0x61, 0x7a, 0x21, 0x70, 0x61, 0x73, 0x73
+                };
+
+
+
+value TSRemoteGuardCreds :: =
+{
+	logonCred {
+		packageName  '4D7953656375726974795061636B616765'H,
+		credBuffer   '4D7920427265617468204973204D792050617373776F7264'H
+	},
+	supplementalCreds {
+		{
+			packageName '416C7465726E617469766553656375726974795061636B616765'H,
+			credBuffer  '4D7920427265617468204973204D79204F746865722050617373776F7264'H
+		},
+		{
+			packageName '50414D'H,
+			credBuffer  '666F6F62617262617A2170617373'H
+		}
+	}
+}
